SUSE support for ceilometer agent-central recipe

+++ recipes/ceilometer-agent-central.rb
@@ -3,6 +3,7 @@
 # Recipe:: ceilometer-agent-central
 #
 # Copyright 2012, AT&T
+# Copyright 2013, SUSE Linux GmbH
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -17,19 +18,30 @@
 # limitations under the License.
 #
 
-class ::Chef::Recipe
-  include ::Openstack
-end
-
 include_recipe "openstack-compute::ceilometer-common"
 
-bindir = '/usr/local/bin'
-ceilometer_conf = node["openstack"]["compute"]["ceilometer"]["conf"]
-conf_switch = "--config-file #{ceilometer_conf}"
+if node["openstack"]["compute"]["platform"]["ceilometer_packages"]
+  node["openstack"]["compute"]["platform"]["ceilometer_packages"]["agent-central"].each do |pkg|
+    package pkg
+  end
 
-service "ceilometer-agent-central" do
-  service_name "ceilometer-agent-central"
-  action [:start]
-  start_command "nohup #{bindir}/ceilometer-agent-central #{conf_switch} &"
-  stop_command "pkill -f ceilometer-agent-central"
+  service node["openstack"]["compute"]["platform"]["ceilometer_services"]["agent-central"] do
+    action :start
+  end
+
+else
+  class ::Chef::Recipe
+    include ::Openstack
+  end
+
+  bindir = '/usr/local/bin'
+  ceilometer_conf = node["openstack"]["compute"]["ceilometer"]["conf"]
+  conf_switch = "--config-file #{ceilometer_conf}"
+
+  service "ceilometer-agent-central" do
+    service_name "ceilometer-agent-central"
+    action [:start]
+    start_command "nohup #{bindir}/ceilometer-agent-central #{conf_switch} &"
+    stop_command "pkill -f ceilometer-agent-central"
+  end
 end
